analytics/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
